refactor(orders-format-util): ordersList 출력 부분 로직 util로 분리
