+++ src/main/java/christmas/util/format/DiscountFormatUtil.java
@@ -1,4 +1,4 @@
-package christmas.util;
+package christmas.util.format;
 
 import static christmas.constants.Symbol.FORMAT_NUMBER;
 import static christmas.constants.Symbol.NEW_LINE;
@@ -8,7 +8,6 @@ import static christmas.constants.ViewMessage.NON_BENEFIT;
 import java.util.Map;
 
 public class DiscountFormatUtil {
-
     public static String formatDiscountResults(Map<String, Integer> discountResults) {
         StringBuilder result = new StringBuilder();
         if (isDiscountNonBenefit(discountResults)) {
